+++ app/routes/scrapping/scrapping.py
@@ -1,11 +1,12 @@
 from flask import Blueprint
+from services.scrapping.index import scrapping_test, scrap
 
 scrapping_bp = Blueprint('scrapping', __name__)
 
-@scrapping_bp.route('/scrapping')
-def scrapping():
-    return "This is the scrapping route."
+@scrapping_bp.route('/test', methods=['GET'])
+def test():
+    return scrapping_test()
 
-@scrapping_bp.route('/scrapping/profile')
-def scrapping_profile():
-    return "This is the scrapping profile route."
+@scrapping_bp.route('/', methods=['POST'])
+def index():
+    return scrap()